refactor and delete logs
